--- aidial_adapter_openai/constant.py
@@ -1,9 +1,0 @@
-import httpx
-
-# connect timeout and total timeout
-DEFAULT_TIMEOUT = httpx.Timeout(600, connect=10)
-
-# Borrowed from openai._constants.DEFAULT_CONNECTION_LIMITS
-DEFAULT_CONNECTION_LIMITS = httpx.Limits(
-    max_connections=1000, max_keepalive_connections=100
-)
