+++ aidial_adapter_openai/gpt.py
@@ -1,5 +1,3 @@
-import logging
-
 from aidial_sdk.exceptions import HTTPException as DialException
 from fastapi.responses import StreamingResponse
 from openai import AsyncStream
@@ -7,17 +5,16 @@ from openai.types.chat.chat_completion import ChatCompletion
 from openai.types.chat.chat_completion_chunk import ChatCompletionChunk
 
 from aidial_adapter_openai.utils.auth import OpenAICreds
-from aidial_adapter_openai.utils.log_config import logger
 from aidial_adapter_openai.utils.parsers import chat_completions_parser
 from aidial_adapter_openai.utils.reflection import call_with_extra_body
 from aidial_adapter_openai.utils.sse_stream import to_openai_sse_stream
-from aidial_adapter_openai.utils.streaming import generate_stream, map_stream
+from aidial_adapter_openai.utils.streaming import (
+    chunk_to_dict,
+    debug_print,
+    generate_stream,
+    map_stream,
+)
 from aidial_adapter_openai.utils.tokens import Tokenizer, discard_messages
-
-
-def debug_print(chunk: dict) -> None:
-    if logger.isEnabledFor(logging.DEBUG):
-        logger.debug(f"chunk: {chunk}")
 
 
 async def gpt_chat_completion(
@@ -59,17 +56,12 @@ async def gpt_chat_completion(
 
     if isinstance(response, AsyncStream):
 
-        def to_dict(chunk: ChatCompletionChunk) -> dict:
-            dict = chunk.to_dict()
-            debug_print(dict)
-            return dict
-
         prompt_tokens = tokenizer.calculate_prompt_tokens(data["messages"])
         return StreamingResponse(
             to_openai_sse_stream(
                 generate_stream(
                     prompt_tokens,
-                    map_stream(to_dict, response),
+                    map_stream(chunk_to_dict, response),
                     tokenizer,
                     deployment_id,
                     discarded_messages,
@@ -81,5 +73,5 @@ async def gpt_chat_completion(
         resp = response.to_dict()
         if discarded_messages is not None:
             resp |= {"statistics": {"discarded_messages": discarded_messages}}
-        debug_print(resp)
+        debug_print("response", resp)
         return resp
