+++ aidial_adapter_openai/mistral.py
@@ -6,16 +6,10 @@ from openai.types.chat.chat_completion import ChatCompletion
 from openai.types.chat.chat_completion_chunk import ChatCompletionChunk
 
 from aidial_adapter_openai.utils.auth import OpenAICreds
-from aidial_adapter_openai.utils.globals import get_http_client
-from aidial_adapter_openai.utils.log_config import logger
+from aidial_adapter_openai.utils.http_client import get_http_client
 from aidial_adapter_openai.utils.reflection import call_with_extra_body
 from aidial_adapter_openai.utils.sse_stream import to_openai_sse_stream
-from aidial_adapter_openai.utils.streaming import map_stream
-
-
-def debug_print(chunk):
-    logger.debug(f"chunk: {chunk}")
-    return chunk
+from aidial_adapter_openai.utils.streaming import chunk_to_dict, map_stream
 
 
 async def chat_completion(
@@ -34,12 +28,7 @@ async def chat_completion(
 
     if isinstance(response, AsyncStream):
         return StreamingResponse(
-            to_openai_sse_stream(
-                map_stream(
-                    debug_print,
-                    map_stream(lambda chunk: chunk.to_dict(), response),
-                )
-            ),
+            to_openai_sse_stream(map_stream(chunk_to_dict, response)),
             media_type="text/event-stream",
         )
     else:
